final code/backend/vision/image_analyzer.py

# backend/vision/image_analyzer.py

import cv2
import os

class ImageAnalyzer:

    def analyze(self, image_path: str):
        """
        Analyze inspection image for safety hazards.
        Currently placeholder logic.
        """
        if not os.path.exists(image_path):
            return {"error": "Image not found"}
        image = cv2.imread(image_path)
        if image is None:
            return {"error": "Invalid image file"}
        # Placeholder hazard detection
        hazards = [
            "Missing helmet",
            "Unsafe machinery"
        ]
        return {
            "image": image_path,
            "hazards": hazards
        }
