final code/backend/requirements.txt
fastapi
uvicorn
openai
python-docx
python-multipart
faiss-cpu
sqlalchemy
passlib
bcrypt
python-jose
python-dotenv
pypdf
requests
beautifulsoup4
prometheus-fastapi-instrumentator
opencv-python-headless
pytest
reportlab
gunicorn
pydantic
opencv-python
psycopg2-binary
redis
celery
PyJWT==2.8.0